server/python/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.4
espn-api==0.31.0 